1.1:
Fixed ScriptTimer.Tick() still being triggered once if TickTimer was
changed to false during Load()
Fixed Tick() timer not starting when TickTimer is set to true after
being set to false.
Added "padding" after key messages from the ServerWrapper to make it
easier on the eyes when reloading scripts.
Added AddDependency(Type t) for scripts to control the Load() order of
scripts.

Added a proof-of-concept MySQL script, could be easily used as a base
for a data storage script.